+++ src/lib/components/collections/collections-list/request/Request.svelte
@@ -17,7 +17,7 @@
   import ModalWrapperV1 from "$lib/components/Modal/Modal.svelte";
   import { notifications } from "$lib/utils/notifications";
   import Button from "$lib/components/buttons/Button.svelte";
-  import RightOption from "$lib/components/right-option/RightOption.svelte";
+  import RightOption from "$lib/components/right-click-menu/RightClickMenuView.svelte";
 
   export let name: string;
   export let id: string;
