+++ src/lib/components/dashboard/rows/Rows.svelte
@@ -1,7 +1,7 @@
 <script>
   import { ThreeDotIcon } from "$lib/assets/app.asset";
   import UserProfileList from "$lib/components/profile/UserProfileList.svelte";
-  import RightOption from "$lib/components/right-option/RightOption.svelte";
+  import RightOption from "$lib/components/right-click-menu/RightClickMenuView.svelte";
 
   export let list;
   export let currOpenedTeamRxDoc;
